Add files via upload
Here I upload the preliminary code.
The result is not very stable yet. i did not perform match parameter tuning.
Also, i used the best test accuracy option to avoid training longer. But it did not help much in some cases. we can try to outperform the best result
of the papers that i included in the zip file. However, it is fine if we do not because our approach is new and we will not limit to classification.
To understand what kind of experiment to be done for classification you can refer to those papers.
I did not include tensorflow code because i did work best only on 40X magnification.
The dataset for this code is on https://web.inf.ufpr.br/vri/databases/breast-cancer-histopathological-database-breakhis/