Readme
### The progress 1 code use vit from 
https://github.com/lucidrains/vit-pytorch

little modification was made.
the dataset is from 
https://web.inf.ufpr.br/vri/databases/breast-cancer-histopathological-database-breakhis/

The performance was not stable. so it is need major parameters tuning.
in the final code that i upload i use top one test accuracy but it does not not change match when the model is train longer before testing.

You can refer to the files i joint with the code for better understanding of the classification experiment and the state of the art result on this dataset.
in only included the best papers.